+++ up42/processing.py
@@ -2,13 +2,13 @@ import abc
 import dataclasses
 import datetime
 import enum
-from typing import ClassVar, List, Optional, TypedDict, Union
+from typing import ClassVar, Iterator, List, Optional, TypedDict, Union
 
 import pystac
 import requests
 import tenacity as tnc
 
-from up42 import base, host
+from up42 import base, host, utils
 
 
 @dataclasses.dataclass(frozen=True)
@@ -30,12 +30,20 @@ class JobStatus(enum.Enum):
     RELEASED = "released"
 
 
+class JobResults(TypedDict, total=False):
+    collection: Optional[str]
+    errors: Optional[List[dict]]
+
+
 class JobMetadata(TypedDict):
-    processID: str  # pylint: disable=invalid-name
-    jobID: str  # pylint: disable=invalid-name
-    accountID: str  # pylint: disable=invalid-name
-    workspaceID: Optional[str]  # pylint: disable=invalid-name
+    # pylint: disable=invalid-name
+    processID: str
+    jobID: str
+    accountID: str
+    workspaceID: Optional[str]
     definition: dict
+    results: Optional[JobResults]
+    creditConsumption: Optional[dict]
     status: str
     created: str
     started: Optional[str]
@@ -47,9 +55,21 @@ class UnfinishedJob(Exception):
     """Job hasn't finished yet with success or failure"""
 
 
+class JobSorting:
+    process_id = utils.SortingField("processID")
+    status = utils.SortingField("status")
+    created = utils.SortingField("created")
+    credits = utils.SortingField("creditConsumption.credits")
+
+
+def _to_datetime(value: Optional[str]):
+    return value and datetime.datetime.fromisoformat(value.rstrip("Z"))
+
+
 @dataclasses.dataclass
 class Job:
     session = base.Session()
+    stac_client = base.StacClient()
     process_id: str
     id: str
     account_id: str
@@ -58,33 +78,40 @@ class Job:
     status: JobStatus
     created: datetime.datetime
     updated: datetime.datetime
+    collection_url: Optional[str] = None
+    errors: Optional[List[ValidationError]] = None
+    credits: Optional[int] = None
     started: Optional[datetime.datetime] = None
     finished: Optional[datetime.datetime] = None
 
-    @staticmethod
-    def __to_datetime(value: Optional[str]):
-        return value and datetime.datetime.fromisoformat(value.rstrip("Z"))
+    @property
+    def collection(self) -> Optional[pystac.Collection]:
+        if self.collection_url is None:
+            return None
+        collection_id = self.collection_url.split("/")[-1]
+        return self.stac_client.get_collection(collection_id)
 
     @staticmethod
     def from_metadata(metadata: JobMetadata) -> "Job":
+        results: JobResults = metadata.get("results") or {}
+        errors = results.get("errors") or []
+        validation_errors = [ValidationError(**error) for error in errors]
+        consumption = metadata.get("creditConsumption") or {}
         return Job(
             process_id=metadata["processID"],
             id=metadata["jobID"],
             account_id=metadata["accountID"],
             workspace_id=metadata["workspaceID"],
+            collection_url=results.get("collection"),
+            errors=validation_errors or None,
+            credits=consumption.get("credits"),
             definition=metadata["definition"],
             status=JobStatus(metadata["status"]),
-            created=Job.__to_datetime(metadata["created"]),
-            started=Job.__to_datetime(metadata["started"]),
-            finished=Job.__to_datetime(metadata["finished"]),
-            updated=Job.__to_datetime(metadata["updated"]),
+            created=_to_datetime(metadata["created"]),
+            started=_to_datetime(metadata["started"]),
+            finished=_to_datetime(metadata["finished"]),
+            updated=_to_datetime(metadata["updated"]),
         )
-
-    @classmethod
-    def get(cls, job_id: str) -> "Job":
-        url = host.endpoint(f"/v2/processing/jobs/{job_id}")
-        metadata = cls.session.get(url).json()
-        return cls.from_metadata(metadata)
 
     def track(self, *, wait: int = 60, retries: int = 60 * 24 * 3):
         @tnc.retry(
@@ -99,11 +126,60 @@ class Job:
             self.updated = job.updated
             self.finished = job.finished
             self.started = job.started
-            # update the results as well
+            self.collection_url = job.collection_url
+            self.errors = job.errors
+            self.credits = job.credits
             if self.status not in [JobStatus.SUCCESSFUL, JobStatus.FAILED]:
                 raise UnfinishedJob
 
         update()
+
+    @classmethod
+    def get(cls, job_id: str) -> "Job":
+        url = host.endpoint(f"/v2/processing/jobs/{job_id}")
+        metadata = cls.session.get(url).json()
+        return cls.from_metadata(metadata)
+
+    @classmethod
+    def all(
+        cls,
+        process_id: Optional[List[str]] = None,
+        workspace_id: Optional[str] = None,
+        status: Optional[List[JobStatus]] = None,
+        min_duration: Optional[int] = None,
+        max_duration: Optional[int] = None,
+        sort_by: Optional[utils.SortingField] = None,
+        *,
+        # used only for performance tuning and testing only
+        page_size: Optional[int] = None,
+    ) -> Iterator["Job"]:
+        query_params = {
+            key: str(value)
+            for key, value in {
+                "workspaceID": workspace_id,
+                "processID": process_id,
+                "status": [entry.value for entry in status] if status else None,
+                "minDuration": min_duration,
+                "maxDuration": max_duration,
+                "limit": page_size,
+                "sort": sort_by,
+            }.items()
+            if value
+        }
+
+        def get_pages():
+            page = cls.session.get(host.endpoint("/v2/processing/jobs"), params=query_params).json()
+            while page:
+                yield page["jobs"]
+                next_page_url = next(
+                    (link["href"] for link in page["links"] if link["rel"] == "next"),
+                    None,
+                )
+                page = next_page_url and cls.session.get(next_page_url).json()
+
+        for page in get_pages():
+            for metadata in page:
+                yield Job.from_metadata(metadata)
 
 
 CostType = Union[int, float, "Cost"]