+++ CHANGELOG.md
@@ -29,6 +29,32 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+## 1.0.4a19
+
+**Jun 13, 2024**
+
+- Added errors and credits retrieval to processing jobs.
+
+## 1.0.4a18
+
+**Jun 13, 2024**
+
+- Deprecate legacy webhook code.
+- Drop long deprecated `Catalog::construct_parameters`.
+
+## 1.0.4a17
+
+**Jun 13, 2024**
+
+- Added `collection` to job class in processing module.
+- Added `StacClient` descriptor to base module.
+
+## 1.0.4a16
+
+**Jun 11, 2024**
+
+- Added job querying to `processing.py`
+
 ## 1.0.4a15
 
 **Jun 11, 2024**